# [jcwgen] perf improvements