@jonathanpeppers 2018-12-12
MSBuild logs: [logs.zip](https://github.com/xamarin/java.interop/files/2673639/logs.zip)