@jonathanpeppers 2018-12-13
Since jcwgen used `ImplementsInterface`, I ran the timings again (three times):
```
Before
1. 1528 ms  GenerateJavaStubs                          1 calls
2. 1488 ms  GenerateJavaStubs                          1 calls 
3. 1522 ms  GenerateJavaStubs                          1 calls 
After
1. 1385 ms  GenerateJavaStubs                          1 calls
2. 1382 ms  GenerateJavaStubs                          1 calls
3. 1424 ms  GenerateJavaStubs                          1 calls 
```
I think it may have improved a little further. Maybe 100ms better now?